+++ R/get_coord_from_string.R
@@ -6,6 +6,7 @@
 #'
 #' @format The returned object is the geographic coordinates using WGS84 (EPSG:4326) in decimal format.
 #' Negative values mean western or southern Hemisphere
+#' @keywords internal
 #'
 #' @examples
 #' \donttest{
@@ -13,7 +14,7 @@
 #'  climate:::get_coord_from_string(txt, pattern = "Latitude")
 #' }
 #'
-get_coord_from_string <- function(txt , pattern = "Longitude") {
+get_coord_from_string <- function(txt, pattern = "Longitude") {
   tt <- gregexpr(pattern, txt)
   start <- tt[[1]][1] + attributes(tt[[1]])$match.length + 1
   tmp <- trimws(substr(txt, start = start, stop = start + 8))