+++ R/hydro_imgw.R
@@ -1,13 +1,13 @@
 #' Hydrological data from IMGW
 #'
-#' Downloading hourly, daily, and monthly hydrological data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection
+#' Downloading hourly, daily, and monthly hydrological data from the measurement stations available in the danepubliczne.imgw.pl collection
 #'
 #' @param interval temporal resolution of the data ("daily" , "monthly", or "semiannual_and_annual")
 #' @param year vector of years (e.g., 1966:2000)
 #' @param coords add coordinates of the stations (logical value TRUE or FALSE)
 #' @param value type of data (can be: state - "H" (default), flow - "Q", or temperature - "T")
-#' @param station vector of hydrological stations danepubliczne.imgw.pl can be name of station CAPITAL LETTERS(character)
-#' It accepts names (characters in CAPITAL LETTERS) or stations' IDs (numeric)
+#' @param station vector of hydrological stations danepubliczne.imgw.pl; can be given as station name with CAPITAL LETTERS (character)
+#' It accepts either names (characters in CAPITAL LETTERS) or stations' IDs (numeric)
 #' @param col_names three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset
 #' @param ... other parameters that may be passed to the 'shortening' function that shortens column names
 #' 
@@ -24,13 +24,13 @@ hydro_imgw <- function(interval, year, coords = FALSE, value = "H", station = NU
 
   if (interval == "daily"){
     # dobowe
-    calosc <- imgw::hydro_daily(year = year, coords = coords, station = station, col_names = col_names, ...)
+    calosc <- hydro_imgw_daily(year = year, coords = coords, station = station, col_names = col_names, ...)
   } else if (interval == "monthly"){
     #miesieczne
-    calosc <- imgw::hydro_monthly(year = year, coords = coords, station = station, col_names = col_names, ...)
+    calosc <- hydro_imgw_monthly(year = year, coords = coords, station = station, col_names = col_names, ...)
   } else if (interval == "semiannual_and_annual"){
     # polroczne_i_roczne
-    calosc <- imgw::hydro_annual(year = year, coords = coords, value = value, station = station, col_names = col_names, ...)
+    calosc <- hydro_imgw_annual(year = year, coords = coords, value = value, station = station, col_names = col_names, ...)
   } else{
     stop("Wrong `interval` value. It should be either 'daily', 'monthly', or 'semiannual_and_annual'.", call. = FALSE)
   }