+++ R/hydro_shortening_imgw.R
@@ -5,6 +5,7 @@
 #' @param data downloaded dataset with original column names
 #' @param col_names three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset
 #' @param remove_duplicates whether to remove duplicated column names (default TRUE - i.e., columns with duplicated names are deleted)
+#' @keywords internal
 #' 
 #' @examples 
 #' \donttest{
@@ -20,7 +21,13 @@
 hydro_shortening_imgw <- function(data, col_names = "short", remove_duplicates = TRUE){
 
   if (col_names != "polish"){
-    abbrev <- imgw::hydro_abbrev
+    abbrev <- climate::imgw_hydro_abbrev
+    
+    # additional workarounds for mac os but not only...
+    abbrev$fullname = gsub(x = abbrev$fullname, pattern="'", replacement = "")
+    abbrev$fullname = gsub(x = abbrev$fullname, pattern="\\^", replacement = "")
+    # end of workaround
+    
     orig_columns <- trimws(gsub("\\s+", " ", colnames(data))) # remove double spaces
 
     matches <- match(orig_columns, abbrev$fullname)