+++ R/meteo_metadata_imgw.R
@@ -1,11 +1,11 @@
 #' Meteorological metadata
 #'
-#' Downloading the description (metadata) to the meteorological data available in the danepubliczne repository.imgw.pl.
+#' Downloading the description (metadata) to the meteorological data available in the danepubliczne.imgw repository.imgw.pl.
 #' By default, the function returns a list or data frame for a selected subset
 #'
 #' @param interval temporal resolution of the data ("hourly", "daily", "monthly")
 #' @param rank rank of station ("synop", "climate", "precip")
-#'
+#' @keywords internal
 #'  
 #' @examples
 #' \donttest{